+++ src/Application/Command/Tip/Event/UpdateHandler.php
@@ -59,7 +59,8 @@ class UpdateHandler
             $command->onPackage,
             $command->onContacts,
             $command->onProgram,
-            $command->onConfirmationPhone
+            $command->onConfirmationPhone,
+            $command->onNetworking
         );
 
         $tip->updateConditions(