+++ src/Application/Query/Happening/Webinar/GetWebinarViewQueryHandler.php
@@ -17,6 +17,7 @@ use Proximum\Vimeet\Application\View\Happening\WebinarSpeakerView;
 use Proximum\Vimeet\Application\View\Happening\Webinar\SpeakerWebinarView;
 use Proximum\Vimeet\Application\View\Happening\Webinar\ViewerWebinarView;
 use Proximum\Vimeet\Application\View\Happening\Webinar\AbstractWebinarView;
+use Proximum\Vimeet\Application\View\Happening\Webinar\WaitingMediaView;
 use Proximum\Vimeet\Domain\Happening\Webinar\IsRecordingAllowed;
 use Proximum\Vimeet\Domain\Model\Happening;
 use Proximum\Vimeet\Domain\Repository\Happening\QuestionRepositoryInterface;
@@ -141,7 +142,8 @@ class GetWebinarViewQueryHandler
                 $happening->getEvent()->getAutoArchiveWebinar(),
                 $questionsCount,
                 $happening->allowWebinarOnHLS(),
-                $viewersCount ?? 0
+                $viewersCount ?? 0,
+                $happening->hasSpeaker($query->getUser())
             );
         }
 
@@ -161,6 +163,10 @@ class GetWebinarViewQueryHandler
             $this->dateTime,
             $timeRemainingInSeconds,
             $happening->getWebinarHeaderImage($query->getLocale()),
+            new WaitingMediaView(
+                $happening->getWebinarWaitingMediaFile($query->getLocale()),
+                $happening->getWebinarWaitingMediaType($query->getLocale())
+            ),
             $liveUrl,
             $happening->isSidebarAllowed(),
             $this->isVideoWebinarAndHappeningIsEnded($happening),