+++ src/Ui/Bundle/EventBundle/Resources/assets/js/components/VideoConference/Webinar.js
@@ -14,6 +14,8 @@ import 'bootstrap/js/popover'; // popover require tooltip
 import Chat from '../_Chat.js';
 import Question from '../_Question.js';
 import NotificationSubscriber from '../_Subscriber';
+import DesktopNotification from './DesktopNotification';
+import Modal from '../Modal';
 
 function Webinar(element, isSpeaker) {
     this.element = element;
@@ -31,6 +33,7 @@ function Webinar(element, isSpeaker) {
     this.newMessageChatCountNotification = element.querySelector('[data-chat-button] span');
     this.newMessageQuestionCountNotification = element.querySelector('[data-questions-button] span');
 
+    this.canDelete = element.hasAttribute('data-chat-can-delete');
 
     if (this.sidebarAllowed) {
         this.shiftWithSidebar = 'shift-with-sidebar';
@@ -52,26 +55,27 @@ function Webinar(element, isSpeaker) {
     this.notificationSubscriberKey = element.getAttribute('data-notifications-subscriber-key');
 
     this.timeRemainingBeforeStart = element.getAttribute('data-time-remaining-before-start');
-    this.timeRemainingBeforeStartMessage = element.getAttribute('data-time-remaining-before-start-message');
+    const modalBeforeStartElement = element.querySelector('[data-modal-warning-before-start]');
+    if (modalBeforeStartElement) {
+        this.modalWarningBeforeStart = new Modal();
+        this.modalWarningBeforeStart.init(modalBeforeStartElement);
+    }
     this.timeRemaining = element.getAttribute('data-time-remaining');
     this.warningRemainingTime = element.getAttribute('data-warning-time-remaining');
 
     if (this.isSpeaker && this.timeRemainingBeforeStart > 0) {
         const startTime = new Date(new Date().getTime() + this.timeRemainingBeforeStart * 1000);
 
-        const timerInterval = setInterval(() => {
-            const remainingTime = Math.round((startTime.getTime() - new Date().getTime()) / 1000);
-
-            if (remainingTime <= 0) {
-                clearInterval(timerInterval);
-                alert(this.timeRemainingBeforeStartMessage);
-            }
-        }, 500);
+        const remainingTime = startTime.getTime() - new Date().getTime();
+        setTimeout(() => {
+            this.modalWarningBeforeStart.show();
+            this.modalWarningBeforeStart.hideAfter(30000);
+        }, remainingTime);
     }
 
     this.chatWaitingMessage = element.getAttribute('data-chat-waiting-message');
     this.userCompleteName = element.getAttribute('data-user-complete-name');
-    this.helperContainer = element.querySelector('.video-helper');
+    this.waitingContainer = element.querySelector('.video-waiting-container');
 
     this.isMobile = 768 > (window.innerWidth || document.documentElement.clientWidth || document.body.clientWidth);
 
@@ -108,6 +112,8 @@ function Webinar(element, isSpeaker) {
     }
 
     this.webinarWaitingMessage = element.querySelector('[data-webinar-waiting-message]');
+    this.webinarWaitingMedia = element.querySelector('[data-webinar-waiting-media]');
+    this.webinarWaitingPlayer = element.querySelector('[data-webinar-waiting-player]');
 
     this.joinButton = element.querySelector('[data-webinar-join-button]');
 
@@ -224,22 +230,34 @@ function Webinar(element, isSpeaker) {
         this.onSettingsValidate.bind(this),
         true
     );
+
+    this.desktopNotificationTitle = element.getAttribute('data-desktop-notification-title');
+    this.desktopNotificationBody = element.getAttribute('data-desktop-notification-body');
+    this.desktopNotification = new DesktopNotification(this.desktopNotificationTitle,this.desktopNotificationBody);
     this.settings.init(true);
 }
 
 Webinar.prototype.onSettingsValidate = function (invisibleMode) {
     this.invisibleMode = invisibleMode;
-    this.join();
-};
+    if (Notification.permission === 'default') {
+        Notification.requestPermission().then(permission => this.join());
+    } else {
+        this.join();
+    }
+}
 
 Webinar.prototype.join = function () {
     this.hideElement(this.joinButton);
 
     if (this.liveUrl) {
-        this.hideElement(this.helperContainer);
+        this.hideElement(this.waitingContainer);
         this.liveVideo();
     } else {
         this.showElement(this.webinarWaitingMessage);
+        this.showElement(this.webinarWaitingMedia);
+        if (this.webinarWaitingPlayer) {
+            this.webinarWaitingPlayer.play();
+        }
     }
 
     this.init();
@@ -262,7 +280,10 @@ Webinar.prototype.init = function () {
 
     if (this.isHls && !this.isSpeaker) {
         this.hlsPlayer = new HlsPlayer(this.hlsVideoElement, () => {
-            this.hideElement(this.helperContainer);
+            if (this.webinarWaitingPlayer) {
+                this.webinarWaitingPlayer.remove();
+            }
+            this.hideElement(this.waitingContainer);
             this.showViewerControls();
             this.layout();
         });
@@ -279,7 +300,10 @@ Webinar.prototype.init = function () {
     this.session = TokboxInstance.initSession(this.apiKey, this.sessionId);
 
     this.session.on('streamCreated', function (event) {
-        this.hideElement(this.helperContainer);
+        if (this.webinarWaitingPlayer) {
+            this.webinarWaitingPlayer.remove();
+        }
+        this.hideElement(this.waitingContainer);
 
         const subscriberManager = new VideoSubscriber(
             this.session,
@@ -582,7 +606,7 @@ Webinar.prototype.showElement = function (element) {
  *  Publish your camera and microphone stream
  */
 Webinar.prototype.publishStream = function () {
-    this.hideElement(this.helperContainer);
+    this.hideElement(this.waitingContainer);
 
     if (this.invisibleMode) {
         return;
@@ -868,15 +892,18 @@ Webinar.prototype.screenshare = function () {
         this.layoutContainer.appendChild(this.screenElement);
         this.session.publish(publisherScreen, this.handlePublishMediaSharing.bind(this));
 
+        this.hasMediaSharing = true;
         this.minimizeAllSubscribers();
         this.maximize(this.screenElement);
         this.layout();
 
         publisherScreen.on('streamCreated', (event) => {
             this.handleStream(event.stream, this.typeScreenShare);
+            this.desktopNotification.showPresent();
         });
         publisherScreen.on('streamDestroyed', (event) => {
             this.handleStopStream(event.stream, this.typeScreenShare);
+            this.desktopNotification.closePresent();
         });
 
         publisherScreen.on('mediaStopped', this.handleStopSharing.bind(this));
@@ -1097,7 +1124,7 @@ Webinar.prototype.addShownChatSubscriber = function () {
         (event) => {
             const payload = JSON.parse(event.data);
 
-            if (payload.action === 'add_chat_message') {
+            if (payload.action === 'add_chat_message' || payload.action === 'delete_chat_message') {
                 this.chat.reload();
             }
 
@@ -1114,6 +1141,10 @@ Webinar.prototype.addHiddenChatSubscriber = function () {
         this.notificationSubscriberKey,
         function (event) {
             const payload = JSON.parse(event.data);
+
+            if (payload.action === 'delete_chat_message') {
+                this.lastSeenChatMessagesCount = Math.max(0, this.lastSeenChatMessagesCount - 1);
+            }
 
             if (payload.action === 'add_chat_message') {
                 const newMessageCount = payload.msg_count - this.lastSeenChatMessagesCount;