+++ src/Ui/Bundle/EventBundle/Resources/assets/js/main.js
@@ -1,3 +1,4 @@
+import '../../../../../../../assets/js/components/Polyfills';
 import '@babel/polyfill';
 import $ from 'jquery';
 import PubSub from 'pubsub-js';