+++ tests/Factory/TipFactory.php
@@ -24,6 +24,7 @@ class TipFactory
     const ON_CONTACTS           = 'onContacts';
     const ON_PROGRAM            = 'onProgram';
     const ON_CONFIRMATION_PHONE = 'onConfirmationPhone';
+    const ON_NETWORKING = 'onNetworking';
 
     const LOCALES = ['fr', 'en'];
 
@@ -48,6 +49,7 @@ class TipFactory
             self::ON_SHEET              => true,
             self::ON_PROGRAM            => true,
             self::ON_CONFIRMATION_PHONE => true,
+            self::ON_NETWORKING         => true,
         ],
         array $locales = self::LOCALES
     ) {
@@ -65,6 +67,7 @@ class TipFactory
             isset($pages[self::ON_CONTACTS]) ? $pages[self::ON_CONTACTS] : false,
             isset($pages[self::ON_PROGRAM]) ? $pages[self::ON_PROGRAM] : false,
             isset($pages[self::ON_CONFIRMATION_PHONE]) ? $pages[self::ON_CONFIRMATION_PHONE] : false,
+            isset($pages[self::ON_NETWORKING]) ? $pages[self::ON_NETWORKING] : false,
             $dateTime
         );