+++ tests/Ui/Bundle/AdminBundle/Controller/Tip/UpdateActionTest.php
@@ -95,6 +95,7 @@ class UpdateActionTest extends TestCase
         $tip->isOnContacts()->willReturn(false);
         $tip->isOnProgram()->willReturn(false);
         $tip->isOnConfirmationPhone()->willReturn(false);
+        $tip->isOnNetworking()->willReturn(false);
         $update = new Update($tip->reveal());
         $form = $this->prophesize(Form::class);
 
@@ -149,6 +150,7 @@ class UpdateActionTest extends TestCase
         $tip->isOnContacts()->willReturn(false);
         $tip->isOnProgram()->willReturn(false);
         $tip->isOnConfirmationPhone()->willReturn(false);
+        $tip->isOnNetworking()->willReturn(false);
         $update = new Update($tip->reveal());
         $form = $this->prophesize(Form::class);